requirements.txt
# Core dependencies
python-dotenv==1.0.1
torch>=2.2.1
huggingface-hub>=0.20.3
transformers>=4.38.2
accelerate==0.25.0
bitsandbytes==0.41.0
einops>=0.7.0
scipy>=1.11.0
sentencepiece>=0.1.99

# LangChain and vector store
langchain>=0.0.350
langchain-community>=0.0.10
chromadb==0.4.24
sentence-transformers==2.5.1

# Document processing
pypdf==4.1.0

# LLM APIs
openai==1.14.0
anthropic>=0.41.0,<1.0.0

# Utilities
pandas==2.2.1
psutil==5.9.8
tqdm==4.66.2
scholarly==1.7.11

# Optional integrations (commented out)
#pyjstor==0.2.0 # fake jstor integration
#wos_client==0.5.0 # fake web of science integration
